Factor out Nunjucks code
