+++ src/index.ts
@@ -1,11 +1,10 @@
 import { createTypstCompiler } from '@myriaddreamin/typst.ts/dist/esm/compiler.mjs';
-import { readFileSync } from 'fs';
 import type { Har } from 'har-format';
-import Nunjucks from 'nunjucks';
 import { type processRequest } from 'trackhar';
-import trackHarTranslationsEn from 'trackhar/i18n/en.json';
 import { generateTyp as generateTypForHar } from './lib/har2pdf';
+import { renderNunjucks } from './lib/nunjucks';
 import { prepareTraffic, type PrepareTrafficOptions } from './lib/traffic';
+import { templates, type SupportedLanguage } from './lib/translations';
 import type { NetworkActivityReport } from './lib/user-network-activity';
 
 export type App = {
@@ -53,13 +52,13 @@ export type ComplaintOptions = {
 export type GenerateOptions =
     | {
           type: 'report' | 'notice';
-          language: 'en';
+          language: SupportedLanguage;
 
           analysis: Analysis;
       }
     | {
           type: 'complaint';
-          language: 'en';
+          language: SupportedLanguage;
 
           /** Data for the initial analysis, that the notice to the controller was based on. */
           initialAnalysis: Analysis;
@@ -67,19 +66,6 @@ export type GenerateOptions =
           analysis: Analysis;
           complaintOptions: ComplaintOptions;
       };
-
-const templates = {
-    en: {
-        report: readFileSync(new URL('../templates/en/report.typ', import.meta.url), 'utf-8'),
-        notice: readFileSync(new URL('../templates/en/notice.typ', import.meta.url), 'utf-8'),
-        complaint: readFileSync(new URL('../templates/en/complaint.typ', import.meta.url), 'utf-8'),
-        style: readFileSync(new URL('../templates/en/style.typ', import.meta.url), 'utf-8'),
-    },
-};
-
-const translations = {
-    en: trackHarTranslationsEn,
-};
 
 export const generate = async (options: GenerateOptions) => {
     // Prepare traffic.
@@ -101,48 +87,18 @@ export const generate = async (options: GenerateOptions) => {
     }
     const { harEntries, trackHarResult, findings } = prepareTraffic(prepareTrafficOptions);
 
-    const nunjucks = Nunjucks.configure({ autoescape: true, throwOnUndefined: true });
-    nunjucks.addFilter('dateFormat', (date: Date | string | undefined, includeTime = true) =>
-        date
-            ? new Date(date).toLocaleString(options.language, {
-                  dateStyle: 'long',
-                  timeStyle: includeTime ? 'long' : undefined,
-              })
-            : undefined
-    );
-    nunjucks.addFilter('timeFormat', (date: Date | string | undefined) =>
-        date ? new Date(date).toLocaleTimeString(options.language) : undefined
-    );
-    // Wrap content in a raw/code block, properly escaping user input.
-    nunjucks.addFilter(
-        'code',
-        (s: string | undefined) =>
-            new Nunjucks.runtime.SafeString(s === undefined ? '' : `\`\`\` ${(s + '').replace(/`/g, '`\u200b')}\`\`\``)
-    );
-    // Convert string from TrackHAR's markup language to Typst.
-    nunjucks.addFilter('trackharMl', (str: string) => str.replace(/\s*\[(https?:\/\/.+?)\]/g, '#footnote[$1]'));
-    // Translate.
-    nunjucks.addGlobal(
-        't',
-        <TScope extends keyof (typeof translations)['en']>(
-            scope: TScope,
-            key: keyof (typeof translations)['en'][TScope] & string
-        ) => (
-            (() => {
-                const translation = translations[options.language][scope][key];
-                if (!translation) throw new Error(`Translation not found: ${scope}::${key}`);
-            })(),
-            translations[options.language][scope][key]
-        )
-    );
-
-    const typSource = nunjucks.renderString(templates[options.language][options.type], {
-        analysis: options.analysis,
-        initialAnalysis: options.type === 'complaint' && options.initialAnalysis,
-        complaintOptions: options.type === 'complaint' && options.complaintOptions,
-        harEntries,
-        trackHarResult,
-        findings,
+    // Render Nunjucks template.
+    const typSource = renderNunjucks({
+        template: templates[options.language][options.type],
+        language: options.language,
+        context: {
+            analysis: options.analysis,
+            initialAnalysis: options.type === 'complaint' && options.initialAnalysis,
+            complaintOptions: options.type === 'complaint' && options.complaintOptions,
+            harEntries,
+            trackHarResult,
+            findings,
+        },
     });
 
     const mainFilePath = '/main.typ';
